Updated Bazaar.mac Added code to check for the minSellPrice going below the vendor value of an item and setting it to the vendor value of the item instead. This is added protection to keep players from inadvertantly creating profit for other characters whom might stuble upon them by using a macro such as buylow.mac
Now when it tries to sell my blue diamond or diamonds it sells it for AT LEAST the vendor value. You have tried to fail, but I won't let you!

Fixed initprice leaving initprice if any item has a price set for it and instead goes to the next item to check it for a value. 
Fixed initprice updating the first item without a price and leaving the loop instead of going to the next item. 
Added color to all /echo commands
Added alias myecho
|||||||/alias /bazecho /echo \ar[\agBazaar.mac\ar]\aw >>> \ao

Updated current version to 3.3, changed the update date to current date.