Add MQ2AutoClaim
Add MQ2AutoClaim - Originally by Dewey2461 Updated by CWTN not to parse macro data, to not need to open the windows, added an INI option to automatically close the popup window after claiming in the event the user has no intention of taking advantage of the heritage crates etc. 
[Settings]
AutoClosePopup=0
Setting this to 1 will cause it to automatically close the Popup. Setting stored in MQ2AutoClaim.ini