Fix crash with /hudlist
refactored SetLast() to get the last Element listed as loaded correctly without crashing.
In HudList() will not display "~None~" if there are no elements to display instead of displaying every section in the INI.

Change Default in GetPrivateProfileString

Changed the GetPrivateProfileString for the character's Last=element,element2
to NULL so that the !strcmp would find that it was in fact "NULL" instead of "Elements" as the previous default had.

remove typdef usages.

CHAR is now char
VOID is now void
PCHAR is now char*
PSTR is now char*

The plugin could do with more cleanup, but just quickly hit the things I spotted and had time to adjust.

Refactor of SetLast() 2.0

rewrite getting last loaded Hud element to use strrchr instead of strtok_s so as to remove looping through each value to get to the last.