time_t adjustments throughout. more castings removed.
where time(nullptr) is used in multiple places within a single function, save the value to a variable and reuse the variable so that when time is used multiple times the same value is used for all entries.
Note: some entries were casted to an (int), such as `MyLast` `MyFirst`, but those are time_t and shouldn't need to be casted, so those casts were removed. 
void DPSMob::DPSEntry::AddDamage(int) has had it's setting of Damage.First moved before it's Damage.Last to match void DPSMob::AddDamage(int)