FindItem/FindItemCount - Duplication of discoveries for augments.
FindItem/FindItemCount iterate through the contents of bags twice when trying to find augments. This causes it to report twice as many augments found. 

How to reproduce the issue: Put an augment in a bag then /echo ${FindItemCount[name of augment here]}

The reason this happens is because bags and augments are iterated through the same way. When checking top level inventory slots, we check for augments, and then we check for bags. But when we check for augments, we don't exclude bags from being checked. 

So for top level inventory checks when checking for augs, I've excluded searching the bags for augs, and instead let that happen when the bags are checked as it also check for augs inside items inside of the bags.